requirements.txt
astroid==1.6.1
Babel==2.5.3
click==6.7
colorama==0.3.9
et-xmlfile==1.0.1
Flask==0.12.2
Flask-Admin==1.5.0
Flask-BabelEx==0.9.3
Flask-Login==0.4.1
Flask-SQLAlchemy==2.3.2
isort==4.3.4
itsdangerous==0.24
jdcal==1.3
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
odfpy==1.3.6
openpyxl==2.5.0
Pillow==5.0.0
pycodestyle==2.3.1
pytz==2018.3
PyYAML==3.12
six==1.11.0
speaklater==1.3
SQLAlchemy==1.2.4
tablib==0.12.1
unicodecsv==0.14.1
Werkzeug==0.14.1
wrapt==1.10.11
WTForms==2.1
xlrd==1.1.0
xlwt==1.3.0
